requirements.txt
python-pptx>=0.6.23
pandas>=2.0.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
pyopenssl>=23.0.0
gunicorn>=21.2.0
waitress>=2.1.2
